Fragment 클래스 분리 후 수정
