+++ Boogimarket/app/src/main/java/com/example/boogimarket/ChatFragment.kt
@@ -1,0 +1,6 @@
+package com.example.boogimarket
+
+import androidx.fragment.app.Fragment
+
+class ChatFragment : Fragment(R.layout.activity_chatroom) {
+}
