--- Boogimarket/app/src/main/java/com/example/boogimarket/Fragments.kt
@@ -1,7 +1,0 @@
-package com.example.boogimarket
-
-import androidx.fragment.app.Fragment
-
-class HomeFragment : Fragment(R.layout.activity_itemlist) {}
-class ChatFragment : Fragment(R.layout.activity_chatroom) {}
-class MyPageFragment : Fragment(R.layout.activity_mypage) {}
