+++ Boogimarket/app/src/main/java/com/example/boogimarket/MainActivity.kt
@@ -11,11 +11,6 @@ import androidx.fragment.app.Fragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
-    private val viewModel by viewModels<MyViewModel>()
-
-    private lateinit var soldCheckBox: CheckBox // 판매된 상품 체크 여부
-    private lateinit var priceRadioGroup: RadioGroup // 가격순 정렬 라디오 버튼 그룹
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_list)
@@ -53,19 +48,6 @@ class MainActivity : AppCompatActivity() {
         }
         return true
 
-    }
-
-    // 판매된 상품 & 가격순 정렬 Filter 함수
-    private fun updateItemsFilters() {
-        val isSold = if (soldCheckBox.isChecked) false else null
-
-        val sortPrice: Boolean? = when (priceRadioGroup.checkedRadioButtonId) {
-            // R.id.lowPrice -> true // 낮은 가격순
-            // R.id.highPrice -> false // 높은 가격순
-            else -> null
-        }
-
-        //viewModel.updateItems(isSold, sortPrice)
     }
 
     // 현재 프래그먼트를 지정된 프래그먼트로 교체
