+++ Boogimarket/app/src/main/java/com/example/boogimarket/MyPageFragment.kt
@@ -1,0 +1,6 @@
+package com.example.boogimarket
+
+import androidx.fragment.app.Fragment
+
+class MyPageFragment : Fragment(R.layout.activity_mypage) {
+}
